% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_attribute}
\alias{get_attribute}
\title{This finds attributes, like output or value, for a particular node by
traversing fails and collecting the named attribute}
\usage{
get_attribute(self, attribute)
}
\arguments{
\item{attribute}{}
}
\value{
a character vector of the node attributes
}
\description{
This finds attributes, like output or value, for a particular node by
traversing fails and collecting the named attribute
}
\examples{
keys <- c("a", "ab", "bab", "bc", "bca", "c", "caa", "abc")
values <- c("foo", "bar", "foo", "bar", "foo", "bar", "foo", "bar")
example_trie <- trie_create(keys, values)
get_attribute(example_trie$b$a$b, "output")
}
