% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trie.R
\name{key_traverse}
\alias{key_traverse}
\title{Traverse the trie to a particular node by full key name.}
\usage{
key_traverse(key, trie)
}
\arguments{
\item{key}{the key to search for}

\item{trie}{the trie to search for}
}
\value{
the trie node that corresponds to the key
}
\description{
Each node's formal name is a single character. You can traverse to the node
of a key in the trie by splitting a key into characters and then traverseing
from parent to child node. This is really just a helper function.
}
\examples{
test <- c("a", "ab", "bab", "bc", "bca", "c", "caa", "abc")
values <- c("foo", "bar", "foo", "bar", "foo", "bar", "foo", "bar")
example_trie <- trie_create(test, value = values)
key_traverse("c", example_trie)
key_traverse("ca", example_trie)
key_traverse("caa", example_trie)
}
