% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trie.R
\name{trie_create}
\alias{trie_create}
\title{Create the finite state automaton.}
\usage{
trie_create(keys, values, to_lower = TRUE, whole_words_only = TRUE)
}
\arguments{
\item{keys}{these are the words that you want to search for in the trie}

\item{values}{these are optional, but if you will need to supply them if you
want to use the replace method.}

\item{to_lower}{are your keys case sensitive? If so, this should be false.}

\item{whole_words_only}{do you want to pad your keys with whitespace to find
only those that are whole words? If so select this should be true}
}
\value{
a reference class object of class node from the data.tree package.
}
\description{
AKA the trie with failure attributes built in. This will need to be run one
time and then can be used for both find and replace functions.
}
\examples{
keys <- c("a", "ab", "bab", "bc", "bca", "c", "caa", "abc")
values <- c("foo", "bar", "foo", "bar", "foo", "bar", "foo", "bar")
example_trie <- trie_create(keys, value = values)
example_trie
example_trie$a$b$fail
}
