% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{trie_find}
\alias{trie_find}
\title{Given a document and a search trie finds occurrences of}
\usage{
trie_find(document, trie, border_chars = c("!", ",", ".", "?", ";", ":",
  "\\n", "\\t"), raw = F)
}
\arguments{
\item{document}{}

\item{trie}{}
}
\value{
a dataframe
}
\description{
Given a document and a search trie finds occurrences of
}
