% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trie.R
\name{trie_insert}
\alias{trie_insert}
\title{Insert new key value pairs into a trie.}
\usage{
trie_insert(trie, key, value)
}
\arguments{
\item{trie}{a trie to update}

\item{key}{the replacement keyword}

\item{value}{the value that will be used to replace keywords when using
trie replace}
}
\value{
an updated trie
}
\description{
This can be used to insert new key-value pairs into an existing trie. It's
also the engine used in trie creation
}
\examples{
test <- c("a", "ab", "bab", "bc", "bca", "c", "caa", "abc")
values <- c("foo", "bar", "foo", "bar", "foo", "bar", "foo", "bar")
example_trie <- trie_create(test, value = values)
trie_insert(example_trie, key = "dab", value = "foo")
#'trie_insert(example_trie, key = "cad", value = "bar")
}
